Fix up some C++ types

+++ unicycler/cpp_function_wrappers.py
@@ -15,7 +15,7 @@ not, see <http://www.gnu.org/licenses/>.
 """
 
 import os
-from ctypes import CDLL, cast, c_char_p, c_int, c_double, c_void_p, c_bool, POINTER
+from ctypes import CDLL, cast, c_char_p, c_int, c_uint, c_ulong, c_double, c_void_p, c_bool, POINTER
 from .misc import quit_with_error
 
 
@@ -215,8 +215,8 @@ def simulate_depths(read_lengths, ref_length, iterations, threads):
 # This function gets the mean and standard deviation of alignments between random sequences.
 C_LIB.multipleSequenceAlignment.argtypes = [POINTER(c_char_p),  # Sequences
                                             POINTER(c_char_p),  # Qualities
-                                            c_int,  # Count
-                                            c_int,  # Bandwidth
+                                            c_ulong,  # Count
+                                            c_uint,  # Bandwidth
                                             c_int,  # Match score
                                             c_int,  # Mismatch score
                                             c_int,  # Gap open score
